Update patch set 6
Patch Set 6:

(7 comments)

A general comment is that we are planning to merge the internal and external codec factories and put all responsibility on the client to create software fallbacks. Maybe we can put the responsibility on the client to create SimulcastEncoderAdapter as well, I don't know the details about this. So a concern I have is that this CL will make it more complicated to move this logic to the codec factories in the future.

deadbeef@ - do you have an opinion about this? The relevant changes are in webrtc/media/engine/webrtcvideoengine2.cc.

Patch-set: 6
Reviewer: Gerrit User 1132136 <1132136@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1117641 <1117641@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>