Update patch set 7
Patch Set 7:

(7 comments)

Acted on your comments @magjed

As a user of the external factory interface on another embedded platform, it would be great to have the fallback moved outside. If and when we do that, we could simply get rid of the software fallback factory and still keep the rest of the simulcast changes in this patch. We simply need to use external encoder directly instead of the SoftwareFallbackEncoderFactory on line 151 in webrtcvideoengine2.cc. We do not use software fallback for the internal factory.

Patch-set: 7
Reviewer: Gerrit User 1176415 <1176415@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>