Update patch set 4
Patch Set 4:

(19 comments)

I've rewritten the factory to use a bunch of helper methods instead of all the "properties" and "requirements" objects.  This makes the code more compact and eliminates all the confusing types.  The possible drawbacks are that it's a bit switch-happy and the information for what configurations are supported is spread over multiple methods.  However, each method is very small and targeted, so it should not be too hard to tell what's going on.

Please take a look and tell me what you think.

Patch-set: 4
Reviewer: Gerrit User 1187699 <1187699@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>