Update patch set 5
Patch Set 5:

(6 comments)

Hi Zeke,

Not sure if you looked through the history of the CL but the different patch sets implement a number of different approaches, e.g. a pure C++ version and a non-singleton version. So there have already been a bunch of discussions about it. I've added some motivations inline in the comments just to clarify the thinking behind it, but if you prefer one of the other approaches I'd love to know. I'll create a new CL and fix a couple of obvious issues that you mentioned, and I can use that to make any other changes as well.

Patch-set: 5