Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (6 comments)
> 
> Hi Zeke,
> 
> Not sure if you looked through the history of the CL but the different patch sets implement a number of different approaches, e.g. a pure C++ version and a non-singleton version. So there have already been a bunch of discussions about it. I've added some motivations inline in the comments just to clarify the thinking behind it, but if you prefer one of the other approaches I'd love to know. I'll create a new CL and fix a couple of obvious issues that you mentioned, and I can use that to make any other changes as well.

Thanks - appreciate it :) I think a C++ version is the correct way to go. This is because otherwise you might as well be calling UIKit directly in the code you want to use it in. The reason we don't want to do that is so that we avoid ObjC++ whenever humanly possible (too many potential issues could happen). In fact, the VideoToolbox classes can now be reverted to .cc files. They were made .mm for a quick fix at some point before RTCUIApplicationIsActive I believe.

The header should be pure C++. You can use the fwd decl macro in RTCMacros.h to declare an NSObject * as a private ivar in the C++ decl. In the ObjC++ impl file, you can define an ObjC interface like @interface RTCUIApplicationObserver along with its definition. It can then have the properties like you have already. The C++ object can create the ObjC one and just access the property it needs.

In the ObjC class impl, there should be one atomic BOOL property that keeps track of state. On init, it can be set to some default - Background might be a safe choice. You can then set observers and then dispatch_async on main to set the ivar to the real value (so there is a short time period before it is truthful - but this is always going to be the case anyway due to thread difference). ObjC property synthesis will give you atomic for free as long as it is decl-ed.

Re: singleton or not - this is an implementation detail separate from C++ header. However - some comments. NSNotificationCenter is designed for this kind of observing, so it's okay to attach multiple observers to it. I think it is cleaner to observe only when you need it, and detach when you don't. Otherwise you have now a library that you can't "unload" after you have started using it. ie in this case, after we started one video encode session, this UIApplication observer lives forever.

Patch-set: 5