Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)
> 
> Have we tried standardizing this? How is it different than packetsDiscarded (https://w3c.github.io/webrtc-stats/#dom-rtcreceivedrtpstreamstats-packetsdiscarded)?
> 
> If you choose to try that route, the way to propose a new stat is by making a GitHub issue here: https://github.com/w3c/webrtc-stats/issues
> 
> Note the guidelines for new stats: https://w3c.github.io/webrtc-stats/#guidelines-for-design-of-stats-objects

That is a very important concern. Thanks for pointing this out.

Reading this
https://w3c.github.io/webrtc-stats/#dom-rtcreceivedrtpstreamstats-packetsdiscarded

The stuff we trying to wire up now is almost the same. So we can basically just use that stat.

However, there is a conceptual difference. The definition of packetsDiscarded says "RTP packets discarded due to packet duplication are not reported in this metric", which we actually wanted to count in this metrics.

But maybe we add a separate stat for packetsDiscarded due to duplication, WDYT?

Patch-set: 2
Reviewer: Gerrit User 1118997 <1118997@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>