Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I'm not sure this approach will work properly. Currently MouseCursorMonitor reports mouse position only when the mouse is on top of the selected window. If the selected window is covered by another window and the mouse cursor is on top of that window then cursor is considered to be outside of the selected window. I don't think you can keep this behavior if the capture monitor is not aware of the selected window.

I have not noticed this. But I believe in general any logic in MouseCursorMonitor can be executed out of it without breaking the output. The benefit is obvious, it needs not to understand SourceId.
Following are several potential solutions,

1. Always returning the window under mouse cursor from MouseCursorMonitor, and add bool DesktopCapturer::IsWindowSourceSelected(WindowId); to check whether the window is covered.

2. Add bool DesktopCapturer::IsPositionVisible(DesktopVector).

3. Check whether the window is covered in DesktopAndCursorComposer.

IMO, the second solution is preferred. Since the check is inside of the DesktopCapturer, the meaning of the SourceId does not matter.

All the solutions decouple MouseCursorMonitor with DesktopCapturer, but leave the logic inside of DesktopAndCursorComposer. The later one needs to know DesktopCapturer anyway.

Or if we allow a small behavior change, i.e. the mouse cursor is only inside of the window if the window is focused:
Add bool DesktopCapturer::IsSelectedSourceFocused().


Meanwhile, adding DesktopRect into DesktopFrame can also help chromoting to transfer video streams from monitors separately.

Patch-set: 3
Reviewer: Gerrit User 1196069 <1196069@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0