Update patch set 10
Patch Set 10:

(40 comments)

I haven't been following the code review closely so I apologize in advance if some of my comments are redundant.

Anyway, I think this is in great shape, and the only things left really are:
1. Adding more comments
2. Renaming/moving around test cases
3. Deciding on the behavior of certain corner cases (there may just be 1 left at this point). I've listed all the ones I've talked about with Peter in the design doc.

Some background: Peter and I had different thoughts for a while about what the corner case behaviors should be. My thinking was "let's make the rules as general as possible; that makes them easier to understand, test and extend with more fields". So I wanted to treat the different parts of the criteria as generically as possible, and have a relatively simple rule of "route a packet to a sink as long as you've narrowed it down to being the only possible sink; don't route a packet if it doesn't match any criteria (due to conflicts) or matches multiple." But what I learned from Peter is that practical considerations had a role in how the BUNDLE algorithm came about; it was a conscious choice that the MID step happens before the SSRC step, which happens before the PT step. Because demuxing on PT or SSRC is more error-prone (with respect to the RTP ecosystem), and when information conflicts, the MID is a more reliable source of information.

So, although I'm not yet sure how to describe the rules in a simple/concise way, there clearly needs to be some "priority" aspect to the rules, like in Peter's doc. It's more complicated than "this packet either matches a criteria or it doesn't", as I was hoping we could get away with.

Patch-set: 10