Check child windows in CroppingWindowCapturerWin
On Windows a window may be covered by its own child window. So this change also
detects child windows by using EnumChildWindow().
This change cannot detect the tooltip or context menu of the child window, so
they may still show up in cropped window frame. See bug for details.

Meanwhile, this change also adds a GetWindowContentRect() to get the content
rectangle of the window. It can help to avoid the case when the border or shadow
of a window to cover the captured window.

Bug: webrtc:8062
Change-Id: I8455a9206d6a1d9da61013ac9debba4d3edae7d8