Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review+1
> 
> LGTM, but is there any documentation on why we're tracking this? I'm not sure what the convention is for adding new histograms.

BlankDetectorDesktopCapturerWrapper: As you may find, it's pretty inefficient and inaccurate to detect whether a frame is blank. But except for reproducing on the virtual machine, we have no idea whether this is really necessary. If not, we should remove it to improve the performance.

FallbackDesktopCapturerWrapper: DirectX capturer fails in lock screen, but we have no idea if it's also impacted in other scenarios. The histogram can give us a clear understanding the usability of DirectX capturer on Windows, and help us calculating a more reliable latency.

Patch-set: 4
Reviewer: Gerrit User 1196069 <1196069@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>