Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Which issues is this change related to? I can't see that it solves any real reported problem.

No real issue, it's more about correctness.

We've wanted to measaure volume of sound, so we've hacked in callback which takes the buffer and do some stuff with it (calculate decibels). But we've find out that the numbers are all wrong, and it's because Buffer's default endian is big, but ARM's native endian is small. Adding this line solves the problem.

So this might save some headache to anyone in the future trying to do the same.

Patch-set: 1
Reviewer: Gerrit User 1240588 <1240588@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0