Update patch set 1
Patch Set 1: Code-Review-1

Thanks but each change in WebRTC should come with a reference to a problem being solved in the existing code base for current users. Adding code to potentially make it easier for some users to do what you are doing in your local environment is not sufficient to motivate the change. It can instead be kept and maintained locally as part of the tests you are running. Feel free to revisit this change if it can be shown to resolve any issued for existing users.

Patch-set: 1
Reviewer: Gerrit User 1001300 <1001300@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1