Update patch set 3
Patch Set 3:

(1 comment)

> Patch Set 2:
> 
> (1 comment)
> 
> I don't think this is the right decision. Surely we do need to fall back on the window capturer in this case, otherwise we'll get another window's border or shadow occluding part of the window we want when we clip the screen capturer result.

I think this is a trade-off,
  - CroppingWindowCapturer can capture more contents, since it's backed by DX capturer. So we should use CroppingWindowCapturer as much as possible.
  - Falling back to use WindowCapturer is because we do not want to expose the content on other windows, or we do not want content on the targeting window to be covered by other windows.
  - Shadow and borders should not be considered as "content" on other windows, because they do not expose any useful information.
  - More important, see the http://crbug.com/741770, task bar has a very large shadow effect, which causes all the fullscreen windows cannot be captured by CroppingWindowCapturer.
  - This breaks lots of new apps, as they are using hardware accelerator.

Considering we are not breaking the principle of "do not expose content on other windows", it should be fine to use content rectangle to determine overlapping. It's really rare for a user to cover one window by using only the border of another one, but do not want to expose the border. On the other hand, it's more common for a user to capture a fullscreen window.

Patch-set: 3
Reviewer: Gerrit User 1196069 <1196069@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>