Update patch set 2
Patch Set 2:

Looks like good changes to me. But when the declaration and definition argument names were homogenized, did you pick the name or did the tool pick the name? It seems like a human ought to be the one to choose which of the two names are better.

Patch-set: 2