Tests for WindowFinder
This change adds several tests for WindowFinder implmenetations.
One of the tests uses ScreenDrawer to create a foreground window or console
window (on Windows) and ensures WindowFinder can return its window_id().
The other ensures WindowFinder always returns kNullWindowId if the |pos| is out
of the screens.

Part of this change duplicates 
https://chromium-review.googlesource.com/c/external/webrtc/+/639590
to make tests pass on X11.
The duplicate logic will be removed after the other change submitted.

Bug: webrtc:7950
Change-Id: I0cc8794e201c2fa042ea8e693434f1b0fa05b47a