Rename IsWindowMinimized to IsWindowNotOnScreen
As discussed in change
https://chromium-review.googlesource.com/c/external/webrtc/+/634043, the name of
IsWindowMinimized() functions is too confusing. IsWindowNotOnScreen() is
preferred, since it describes the behavior of the functions more accurately.

This change does not flip the default behavior of these functions to avoid a
behavior change. TODO has been added; and the flipping will happen in a
following change.

Bug: chromium:758554
Change-Id: I009c0fa57142756e5c83f76b2a3561253db1b67f