Move SDP m= line matching from BaseChannel to WebRtcSession
This is part of the work towards implementing Unified Plan.

The logic for correlating m= lines to channels is changing in
Unified Plan. Moving this logic to WebRtcSession means that we do
not need to add a flag to BaseChannel to indicate which logic it
should use (i.e., Plan B vs. Unified Plan) and can keep those
details in WebRtcSession.

Bug: webrtc:8183
Change-Id: I729da73ece01fd20f45e82f8956a02c4cad2469e